refactor: rename exchange.types
