+++ freqtrade/exchange/exchange_types.py
