+++ freqtrade/plugins/pairlist/OffsetFilter.py
@@ -6,7 +6,7 @@ import logging
 from typing import Dict, List
 
 from freqtrade.exceptions import OperationalException
-from freqtrade.exchange.types import Tickers
+from freqtrade.exchange.exchange_types import Tickers
 from freqtrade.plugins.pairlist.IPairList import IPairList, PairlistParameter, SupportsBacktesting
