+++ freqtrade/plugins/pairlist/PerformanceFilter.py
@@ -7,7 +7,7 @@ from typing import Dict, List
 
 import pandas as pd
 
-from freqtrade.exchange.types import Tickers
+from freqtrade.exchange.exchange_types import Tickers
 from freqtrade.persistence import Trade
 from freqtrade.plugins.pairlist.IPairList import IPairList, PairlistParameter, SupportsBacktesting
