+++ freqtrade/plugins/pairlist/ProducerPairList.py
@@ -8,7 +8,7 @@ import logging
 from typing import Dict, List, Optional
 
 from freqtrade.exceptions import OperationalException
-from freqtrade.exchange.types import Tickers
+from freqtrade.exchange.exchange_types import Tickers
 from freqtrade.plugins.pairlist.IPairList import IPairList, PairlistParameter, SupportsBacktesting
